Split leadership.rs into smaller modules
This commit moves the LeaderState and the SelfProposer into their
own modules in order to make the leadership module easier to reason
about.
